feat: Move CliProgram to a deno-only js-runtime export FileSystemProgramResolver, to use in integration tests for mutlifile patterns.
