+++ packages/patterns/counter.tsx
@@ -1,45 +1,9 @@
 /// <cts-enable />
-import {
-  Cell,
-  Default,
-  derive,
-  h,
-  handler,
-  NAME,
-  Opaque,
-  OpaqueRef,
-  recipe,
-  str,
-  UI,
-} from "commontools";
+import { Default, h, NAME, recipe, str, UI } from "commontools";
+import { decrement, increment, nth, previous } from "./counter-handlers.ts";
 
 interface RecipeState {
   value: Default<number, 0>;
-}
-
-const increment = handler<unknown, { value: Cell<number> }>((_, state) => {
-  state.value.set(state.value.get() + 1);
-});
-
-const decrement = handler((_, state: { value: Cell<number> }) => {
-  state.value.set(state.value.get() - 1);
-});
-
-function previous(value: number) {
-  return value - 1;
-}
-
-function nth(value: number) {
-  if (value === 1) {
-    return "1st";
-  }
-  if (value === 2) {
-    return "2nd";
-  }
-  if (value === 3) {
-    return "3rd";
-  }
-  return `${value}th`;
 }
 
 export default recipe<RecipeState>("Counter", (state) => {
